Extension to background
A few minor changes have occured, with the background being extended further. If the character goes left, a black panel will be visible, else the system works ahead. Minor issues with the clouds, but otherwise, all clear. Future developments are to be made with an automatic background, so the user will not be able to go backwards (potential). Updates, to be produced in the next 12 hours.